+++ python/python/Ice/PropertiesAdminUpdateCallback.py
@@ -1,0 +1,11 @@
+# -*- coding: utf-8 -*-
+#
+# Copyright (c) ZeroC, Inc. All rights reserved.
+#
+
+class PropertiesAdminUpdateCallback(object):
+    """Callback class to get notifications of property updates passed
+    through the Properties admin facet"""
+
+    def updated(self, props):
+        pass
